+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [4.0.4-beta]
 
-Bug fix.
+Repair outdated config files on game load.
 
 ### Fixed
-- Update to fixed Poly Mixin version for Fabric 1.20.1
+- Update to fixed Poly Mixin version for Fabric 1.20.1 (less unstable)
 - Fix floating torches (or others using support shape methods)
+- Fix invalid config file crash
 
 ## [4.0.3-beta]